Added the ability to save/load and start a new game. Improved UI to be more intuitive. Adjusted the scaling of various upgrades.

+++ IdleGame.Domain/Resource.cs
@@ -5,7 +5,7 @@ public class Resource : IResource
     public string Name { get; }
     public long Quantity { get; private set; }
 
-    public Resource(string name, int quantity)
+    public Resource(string name, long quantity)
     {
         Name = name?.Trim() ?? throw new ArgumentNullException(nameof(name));
         Quantity = quantity;
